# api/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EIAM API Documentation",
    "description": "基于 AST 与反射契约混合自适应生成的 OpenAPI 3.0 规范，支持导入 Apifox / Postman 进行全量接口联调。",
    "version": "1.0.0"
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/group/create": {
      "post": {
        "tags": [
          "用户分组"
        ],
        "summary": "创建用户组",
        "description": "权限标识: add",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupRequest"
              },
              "example": {
                "code": "admin",
                "desc": "相关业务描述信息",
                "name": "示例名称"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/group/delete/{id}": {
      "delete": {
        "tags": [
          "用户分组"
        ],
        "summary": "删除用户组",
        "description": "权限标识: delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "唯一数字标识 ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/group/detail/{code}": {
      "get": {
        "tags": [
          "用户分组"
        ],
        "summary": "用户组详情",
        "description": "权限标识: get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "description": "唯一标识编码 Code",
            "schema": {
              "type": "string",
              "example": "admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/group/list": {
      "post": {
        "tags": [
          "用户分组"
        ],
        "summary": "用户组列表",
        "description": "权限标识: view",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListGroupRequest"
              },
              "example": {
                "limit": 10,
                "offset": 0
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/group/list/attached/role": {
      "post": {
        "tags": [
          "用户分组"
        ],
        "summary": "查询角色关联分组",
        "description": "权限标识: view_role_groups",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListAttachedRoleGroupsRequest"
              },
              "example": {
                "keyword": "search_key",
                "limit": 10,
                "offset": 0,
                "role_code": "admin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/group/list/attached/user": {
      "post": {
        "tags": [
          "用户分组"
        ],
        "summary": "查询用户所属分组",
        "description": "权限标识: view_user_groups",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListAttachedUserGroupsRequest"
              },
              "example": {
                "keyword": "search_key",
                "limit": 10,
                "offset": 0,
                "user_id": 1,
                "username": "示例名称"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/group/members": {
      "post": {
        "tags": [
          "用户分组"
        ],
        "summary": "组成员列表",
        "description": "权限标识: members",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListMembersRequest"
              },
              "example": {
                "group_code": "admin",
                "keyword": "search_key",
                "limit": 10,
                "offset": 0
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/group/members/assign": {
      "post": {
        "tags": [
          "用户分组"
        ],
        "summary": "分配组成员",
        "description": "权限标识: assign_members",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignMembersRequest"
              },
              "example": {
                "group_code": "admin",
                "usernames": [
                  "admin",
                  "guest"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/group/members/remove": {
      "post": {
        "tags": [
          "用户分组"
        ],
        "summary": "移除组成员",
        "description": "权限标识: remove_members",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveMembersRequest"
              },
              "example": {
                "group_code": "admin",
                "usernames": [
                  "admin",
                  "guest"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/group/role/assign": {
      "post": {
        "tags": [
          "用户分组"
        ],
        "summary": "用户组分配角色",
        "description": "权限标识: assign_role",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleRequest"
              },
              "example": {
                "role_code": "admin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/group/role/remove": {
      "post": {
        "tags": [
          "用户分组"
        ],
        "summary": "用户组解绑角色",
        "description": "权限标识: remove_role",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveRoleRequest"
              },
              "example": {
                "group_code": "admin",
                "role_code": "admin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/group/roles/{code}": {
      "get": {
        "tags": [
          "用户分组"
        ],
        "summary": "查看用户组角色",
        "description": "权限标识: roles",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "description": "唯一标识编码 Code",
            "schema": {
              "type": "string",
              "example": "admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
    "/api/group/update": {
      "post": {
        "tags": [
          "用户分组"
        ],
        "summary": "修改用户组",
        "description": "权限标识: edit",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupRequest"
              },
              "example": {
                "desc": "相关业务描述信息",
                "id": 1,
                "name": "示例名称"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功"
          }
        }
      }
    },
...
  },
...
  "components": {
    "schemas": {
      "AssignMembersRequest": {
        "type": "object",
        "properties": {
          "group_code": {
            "type": "string",
            "example": "admin"
          },
          "usernames": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "admin",
              "guest"
            ]
          }
        },
        "required": [
          "group_code",
          "usernames"
        ],
        "example": {
          "group_code": "admin",
          "usernames": [
            "admin",
            "guest"
          ]
        }
      },
      "AssignRoleRequest": {
        "type": "object",
        "properties": {
          "role_code": {
            "type": "string",
            "example": "admin"
          }
        },
        "required": [
          "role_code"
        ],
        "example": {
          "role_code": "admin"
        }
      },
...
      "CreateGroupRequest": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "admin"
          },
          "desc": {
            "type": "string",
            "example": "相关业务描述信息"
          },
          "name": {
            "type": "string",
            "example": "示例名称"
          }
        },
        "required": [
          "name",
          "code"
        ],
        "example": {
          "code": "admin",
          "desc": "相关业务描述信息",
          "name": "示例名称"
        }
      },
...
      "ListAttachedRoleGroupsRequest": {
        "type": "object",
        "properties": {
          "keyword": {
            "type": "string",
            "example": "search_key"
          },
          "limit": {
            "type": "integer",
            "example": 10
          },
          "offset": {
            "type": "integer",
            "example": 0
          },
          "role_code": {
            "type": "string",
            "example": "admin"
          }
        },
        "required": [
          "role_code"
        ],
        "example": {
          "keyword": "search_key",
          "limit": 10,
          "offset": 0,
          "role_code": "admin"
        }
      },
      "ListAttachedUserGroupsRequest": {
        "type": "object",
        "properties": {
          "keyword": {
            "type": "string",
            "example": "search_key"
          },
          "limit": {
            "type": "integer",
            "example": 10
          },
          "offset": {
            "type": "integer",
            "example": 0
          },
          "user_id": {
            "type": "integer",
            "example": 1
          },
          "username": {
            "type": "string",
            "example": "示例名称"
          }
        },
        "example": {
          "keyword": "search_key",
          "limit": 10,
          "offset": 0,
          "user_id": 1,
          "username": "示例名称"
        }
      },
      "ListGroupRequest": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "integer",
            "example": 10
          },
          "offset": {
            "type": "integer",
            "example": 0
          }
        },
        "example": {
          "limit": 10,
          "offset": 0
        }
      },
      "ListMembersRequest": {
        "type": "object",
        "properties": {
          "group_code": {
            "type": "string",
            "example": "admin"
          },
          "keyword": {
            "type": "string",
            "example": "search_key"
          },
          "limit": {
            "type": "integer",
            "example": 10
          },
          "offset": {
            "type": "integer",
            "example": 0
          }
        },
        "required": [
          "group_code"
        ],
        "example": {
          "group_code": "admin",
          "keyword": "search_key",
          "limit": 10,
          "offset": 0
        }
      },
...
      "RemoveMembersRequest": {
        "type": "object",
        "properties": {
          "group_code": {
            "type": "string",
            "example": "admin"
          },
          "usernames": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "admin",
              "guest"
            ]
          }
        },
        "required": [
          "group_code",
          "usernames"
        ],
        "example": {
          "group_code": "admin",
          "usernames": [
            "admin",
            "guest"
          ]
        }
      },
      "RemoveRoleRequest": {
        "type": "object",
        "properties": {
          "group_code": {
            "type": "string",
            "example": "admin"
          },
          "role_code": {
            "type": "string",
            "example": "admin"
          }
        },
        "required": [
          "group_code",
          "role_code"
        ],
        "example": {
          "group_code": "admin",
          "role_code": "admin"
        }
      },
...
      "UpdateGroupRequest": {
        "type": "object",
        "properties": {
          "desc": {
            "type": "string",
            "example": "相关业务描述信息"
          },
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "示例名称"
          }
        },
        "required": [
          "id"
        ],
        "example": {
          "desc": "相关业务描述信息",
          "id": 1,
          "name": "示例名称"
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "请输入登录后获取的 JWT Bearer Token 进行接口调用鉴权"
      }
    }
  }
}